Moved composable to individual file
